requirements.txt
Django==4.0.4
djangorestframework==3.13.1
redis==4.3.1
aiohttp==3.8.1
requests==2.27.1
